requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
importlib-metadata==8.4.0
importlib-resources==6.4.4
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
typing-extensions==4.12.2
werkzeug==3.0.4
zipp==3.20.1
